# openapi.yaml
openapi: 3.0.3
info:
  title: SafeGirl Backend API
  description: |
    Women's safety reporting platform with encrypted reports, blockchain verification,
    non-blocking panic alerts, and report sharing with access control.

    **Key Features:**
    - Phone-only OTP authentication (no passwords)
    - AES-256-GCM encryption for report data
    - IPFS + Polygon blockchain for immutable report records
    - Non-blocking panic alerts with SMS notifications
    - Report access control and sharing
    - Key recovery with PIN backup
  version: 1.0.0
  contact:
    name: SafeGirl Team
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.safegirl.example.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from signup/login verify endpoints
  schemas:
...
    ReportStatusResponse:
      type: object
      required: [reportId, status]
      properties:
        reportId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed]
          example: "confirmed"
        ipfsHash:
          type: string
          nullable: true
        txHash:
          type: string
          nullable: true
        blockNumber:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
          nullable: true
...
    SearchReportsResponse:
      type: object
      required: [reports, total]
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/ReportStatusResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    ReportStatsResponse:
      type: object
      required: [totalReports, pendingReports, confirmedReports, failedReports]
      properties:
        totalReports:
          type: integer
        pendingReports:
          type: integer
        confirmedReports:
          type: integer
        failedReports:
          type: integer
        byMonth:
          type: object
          description: Reports submitted per month
paths:
...
  /api/search/reports:
    get:
      summary: Search and filter reports
      operationId: searchReports
      tags: [Search]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, failed]
        - name: createdAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchReportsResponse'
  /api/search/stats:
    get:
      summary: Get report statistics
      operationId: getReportStats
      tags: [Search]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Report statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportStatsResponse'